mainContent

+++ src/main/java/com/cccr/armtong/vo/TemperatureBasicVo.java
@@ -4,13 +4,13 @@ public class TemperatureBasicVo {
     
     private int temperature_idx;
     private int member_idx;
-    private double temperature_tem;
+    private Float temperature_tem;
     private String temperature_date;
     private String temperature_location;
 
     public TemperatureBasicVo() {}
 
-    public TemperatureBasicVo(int temperature_idx, int member_idx, double temperature_tem, String temperature_date, String temperature_location){
+    public TemperatureBasicVo(int temperature_idx, int member_idx, Float temperature_tem, String temperature_date, String temperature_location){
         super();
         this.temperature_idx = temperature_idx;
         this.member_idx = member_idx;
@@ -35,11 +35,11 @@ public class TemperatureBasicVo {
         this.member_idx = member_idx;
     }
 
-    public double getTemperature_tem(){
+    public Float getTemperature_tem(){
         return temperature_tem;
     }
 
-    public void setTemperature_tem(double temperature_tem){
+    public void setTemperature_tem(Float temperature_tem){
         this.temperature_tem = temperature_tem;
     }
